+++ srcs/Config.cpp
@@ -45,8 +45,8 @@ void	CONF::Config::parseServers(string file)
 vector<string>	CONF::Config::chopString(string str)
 {
 	vector<string> v;
-    int		start = 0;
-    int		end = 0;
+    size_t	start = 0;
+    size_t	end = 0;
 	string	delimiter = " \t;";
 
 	while ((start = str.find_first_not_of(delimiter, end)) != string::npos)
@@ -61,7 +61,7 @@ vector<string>	CONF::Config::chopString(string str)
 
 void	CONF::Config::showInfo()
 {
-	for (int i = 0; i < this->servers.size(); i++)
+	for (size_t i = 0; i < this->servers.size(); i++)
 		this->servers[i].printInfo();
 	cout << "Total number of servers: " << this->servers.size() << endl;
 }
